refactor: parser replaced with my lib

+++ lib/configuration/configuration.h
@@ -1,6 +1,7 @@
 #ifndef CONFIGURATION
 #define CONFIGURATION
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -9,8 +10,8 @@
 struct Configuration {
     std::vector<City> cities;
     std::string api_key;
-    int days;
-    int frequency;
+    int64_t days;
+    int64_t frequency;
 };
 
 #endif  // !CONFIGURATION
